# Improve performance of share indicators