## Description
We've improved the performance of share indicators when loading resource tables as well as when adding or removing shares.

In detail:

* Accessing a shared folder: The share indicators now get loaded initially with the resources. The store then gets updated once, including all indicators. Before, the store was updated after the initial load. Also, this happened for each resource, resulting in a re-render of each table row.
* Creating shares & public links: Instead of loading share indicators for all resources in the store, we now only load those for the current file (and files inside a folder).

## Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
- https://github.com/owncloud/web/issues/7038

## Performance measurements

Tests were done with a shared folder including 800 files. Note that the measured time is generally longer when the measurement tool is running. This means the tool is fine for a before- & after-comparison, but not for evaluating the total time as it would be much faster without the measurement.

**Accessing the folder:**

Without optimization (~ 8,5 s):

![Pasted Graphic 1](https://user-images.githubusercontent.com/50302941/176133481-b313d262-45df-4ba0-952f-4b9151aedd9d.png)

With optimization (~ 6,5 s):

![Pasted Graphic](https://user-images.githubusercontent.com/50302941/176133408-77d89629-b6ea-4c77-878b-ed421b6a53b5.png)

**Creating a public link for a resource within the folder:**

Without optimization (~ 11 s):

![image](https://user-images.githubusercontent.com/50302941/176176302-65b150aa-3c83-47b7-b932-2252bdbb7d85.png)

With optimization (~ 6,5 s):

![image](https://user-images.githubusercontent.com/50302941/176175336-f2402bd6-3527-47c7-ab30-3754d82b9b31.png)


## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [x] Technical debt
- [ ] Tests